nix: Add cl-duckdb, make bin/permo wrapper script
Add cl-duckdb as a lisp dependency. Source from my Github PR branch to
nixpkgs because this is a new package outside of quicklisp.

Make bin/permo a wrapper shellscript instead of an executable SBCL
image. The wrapper script sets LD_LIBRARY_PATH and runs sbcl with the
permo core file.

(It didn't work to add the libraries to the executable
image's ELF rpath with patchelf. That patching seems to break sbcl's
executable image magic. Hence switching to a wrapper script.)

Also added "parachute" as a Lisp dependency. This change just came along
for the ride. YOLO!